Add files via upload
# UEFI RAM and BIOS Information

This UEFI application retrieves and prints information about RAM and BIOS using the UEFI protocols and libraries.

## Table of Contents
- [Introduction](#introduction)
- [Prerequisites](#prerequisites)
- [Usage](#usage)
- [Building](#building)
- [License](#license)

## Introduction

This UEFI application demonstrates the retrieval and printing of RAM size information and BIOS details, including the BIOS vendor, version, release date, and content.

## Prerequisites

Before building and running this application, make sure you have the necessary UEFI development environment set up. Ensure that you have a compatible UEFI firmware, and a toolchain capable of building UEFI applications.

## Usage

Compile the application and load it onto a UEFI-compatible system. When executed, it will print the RAM size in gigabytes and information about the BIOS.

## Building

To build the application, use the following steps:

1. Set up your UEFI development environment.
2. Place this source code in the appropriate directory.
3. Use the provided build tools to compile the application.

For example, using the EDK2 build system: